Fixed minor issue with conversion to int and such

+++ input/test.c
@@ -2,6 +2,6 @@
 * My program
 */
 int x = 5*(3/10 + 9/10);
-float y = x*2/( 2+1 * 2/3 +x) +8 * (8/4);
-float result = x + y; //calculate the result
+int y = x*2/( 2+1 * 2/3 +x) +8 * (8/4);
+int result = x + y; //calculate the result
 printf(result); //show the result
